parsing: prepare for the 'the lexer hack'
This is needed in order to parse type names/variables correctly:
for example, if we see the statement

{ a * b ; }

it could mean that 'b' is being declared as a pointer to 'a', or that
the expression 'a * b' is supposed to be executed for its side-effects.
In order to decide between the two, we need to know that 'a' is a type
name.